+++ crates/matrix-sdk/tests/integration/room/beacon/mod.rs
@@ -1,8 +1,8 @@
 use std::time::{Duration, UNIX_EPOCH};
 
-use matrix_sdk::{config::SyncSettings, instant::SystemTime};
+use matrix_sdk::config::SyncSettings;
 use matrix_sdk_test::{async_test, mocks::mock_encryption_state, test_json, DEFAULT_TEST_ROOM_ID};
-use ruma::event_id;
+use ruma::{event_id, time::SystemTime};
 use serde_json::json;
 use wiremock::{
     matchers::{body_partial_json, header, method, path_regex},