+++ crates/matrix-sdk-base/CHANGELOG.md
@@ -15,6 +15,7 @@
   - They both take a `MediaThumbnailSettings` instead of `MediaThumbnailSize`.
 - The `StateStore` methods to access data in the media cache where moved to a separate
   `EventCacheStore` trait.
+- The `instant` module was removed, use the `ruma::time` module instead.
 
 # 0.7.0